Fix Float32Array size
Before that patch, it creates a wrong `Float32Array`, based on the size in bytes (`count * stride`). 

That sounds wrong, since `Float32Array` expect the size in `Float32` (4 bytes). This patch replaces it to `count * (stride / 4)`. The size (in bytes) can be compared using `count * stride === data.byteLength` which will return `true`.

-----

That change improves the performance, on my machine from ~150_000 to ~350_000.